default background
Tried to do a tiled background with CSS but wasn't able to get quite
what I wanted, though rather near to it. Simple tiles of any fixed size would've
been fairly easy, as would've doing 3 tiles of a color and one tile of the other color,
but using relative dimensions and wanting the equal amount of tiles with the matching colors
gets quite a bit more complicated. Need to think about this a bit more - I could
probably circumvent the problem in some way. But at least for now this is fine and
anyway I'm not even convinced the tiles are what I'll want in the end.