Make the mdx pages simpler.
No need for the folders and wrapping all of the mdx pages in the layout in a separate
component, but instead configure all of them to be wrapped within the layout component
by default.

Previously gatsby also actually created separate pages for both the index file and the mdx
file within the same folder, which was kind of messy.